refactor(reborn): isolate webui auth concerns
