+++ crates/ironclaw_reborn_composition/CLAUDE.md
@@ -112,10 +112,11 @@ Inbound order (outer → inner → handler):
    `config.tenant_id` plus the authentication result's `UserId`, and a
    request-scoped `WebUiV2Capabilities` extension from the same
    authentication result.
-   When `openai-compat-beta` is enabled, the same verified bearer result also
-   inserts an `OpenAiCompatAuthenticatedCaller` extension with tenant-scoped
-   verified auth evidence for protected OpenAI-compatible route mounts; route
-   crates must not mint this evidence.
+   When `openai-compat-beta` is enabled, `openai_compat_serve` wraps protected
+   OpenAI-compatible route mounts with middleware that derives an
+   `OpenAiCompatAuthenticatedCaller` from the verified `WebUiAuthenticatedCaller`
+   and tenant-scoped verified auth evidence; generic HTTP kit and route crates
+   must not mint this product-specific evidence.
 8. **Descriptor-driven per-route rate limit**
    (`webui_rate_limit::enforce_rate_limit`) — reads
    `ironclaw_webui_v2::webui_v2_routes()` plus mounted product-auth
