+++ crates/ironclaw_reborn_http_kit/src/serve.rs
@@ -39,16 +39,14 @@ use std::sync::Arc;
 
 use axum::{
     Json, Router,
-    extract::{Request, State},
     http::{HeaderName, HeaderValue, Method, StatusCode, header},
-    middleware::{self, Next},
-    response::{IntoResponse, Response},
+    middleware,
     routing::get,
 };
 use ironclaw_host_api::ingress::IngressRouteDescriptor;
-use ironclaw_host_api::{AgentId, ProjectId, TenantId, UserId};
+use ironclaw_host_api::{AgentId, ProjectId, TenantId};
 use ironclaw_webui_v2::{
-    DEFAULT_SSE_MAX_CONCURRENT_PER_CALLER, WebUiV2Capabilities, WebUiV2RouteOptions, WebUiV2State,
+    DEFAULT_SSE_MAX_CONCURRENT_PER_CALLER, WebUiV2RouteOptions, WebUiV2State,
     is_webui_v2_operator_webui_config_route_id, webui_v2_router_with_options,
 };
 use tower_http::catch_panic::CatchPanicLayer;
@@ -56,13 +54,11 @@ use tower_http::cors::{AllowHeaders, CorsLayer};
 use tower_http::limit::RequestBodyLimitLayer;
 use tower_http::set_header::SetResponseHeaderLayer;
 
+use crate::auth::{AuthLayerState, WebuiAuthenticator, authenticate_request};
 use crate::body_limit::{build_body_limit_state, enforce_body_limit};
-use crate::operator_auth::{
-    OperatorWebuiConfigRouteState, build_operator_webui_config_route_state,
-};
+use crate::operator_auth::build_operator_webui_config_route_state;
 use crate::rate_limit::{build_rate_limit_state, enforce_rate_limit};
 use crate::ws_origin::{build_websocket_origin_state, enforce_websocket_origin};
-use ironclaw_product_workflow::WebUiAuthenticatedCaller;
 use serde::Serialize;
 
 /// Default per-request body limit (14 MiB) — sized to cover ~10 MiB of
@@ -79,79 +75,6 @@ pub(crate) const DEFAULT_WEBUI_CSP: &str =
     "default-src 'self'; object-src 'none'; frame-ancestors 'none'; base-uri 'self'";
 
 const REBORN_HEALTH_PATH: &str = "/api/health";
-
-/// Authentication contract the Reborn binary supplies. The composition
-/// layer is intentionally agnostic about WHERE bearer tokens come from
-/// — env vars, the host's `SecretStore`, OIDC JWTs verified by the
-/// caller — so the same `webui_v2_app` works for the CLI binary and
-/// for any future ingress fronting the same routes.
-///
-/// Implementations return `Some(UserId)` on success and `None` to
-/// reject. Concrete failure reasons stay inside the implementation
-/// (the gateway emits a generic 401), per the
-/// `docs/reborn/how-to-port-channel-to-reborn.md` Path A guidance that
-/// auth evidence is host-owned and never leaks to clients.
-#[async_trait::async_trait]
-pub trait WebuiAuthenticator: Send + Sync + 'static {
-    /// Authenticate a bearer and return the caller identity plus capabilities
-    /// that apply to this exact token.
-    async fn authenticate(&self, token: &str) -> Option<WebuiAuthentication>;
-
-    /// Whether bearer tokens accepted by this authenticator represent a
-    /// single trusted operator. Operator-wide WebUI config routes mutate
-    /// shared host configuration such as provider catalogs, secrets, active
-    /// models, or Slack channel routes, so host composition only mounts them
-    /// for authenticators that explicitly opt in.
-    fn mounts_operator_webui_config_routes(&self) -> bool {
-        #[allow(deprecated)]
-        self.allows_operator_webui_config()
-    }
-
-    #[deprecated(
-        since = "0.1.0",
-        note = "Use `mounts_operator_webui_config_routes`; this asks about route availability, not per-token capability."
-    )]
-    fn allows_operator_webui_config(&self) -> bool {
-        #[allow(deprecated)]
-        self.allows_operator_llm_config()
-    }
-
-    #[deprecated(
-        since = "0.1.0",
-        note = "Use `mounts_operator_webui_config_routes`; this asks about route availability, not per-token capability."
-    )]
-    fn allows_operator_llm_config(&self) -> bool {
-        false
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct WebuiAuthentication {
-    pub user_id: UserId,
-    pub capabilities: WebUiV2Capabilities,
-}
-
-impl WebuiAuthentication {
-    pub fn new(user_id: UserId, capabilities: WebUiV2Capabilities) -> Self {
-        Self {
-            user_id,
-            capabilities,
-        }
-    }
-
-    pub fn user(user_id: UserId) -> Self {
-        Self::new(user_id, WebUiV2Capabilities::default())
-    }
-
-    pub fn operator(user_id: UserId) -> Self {
-        Self::new(
-            user_id,
-            WebUiV2Capabilities {
-                operator_webui_config: true,
-            },
-        )
-    }
-}
 
 /// Deployment gate for the Reborn Projects WebUI surface. Read once here
 /// (host-owned config — per the composition crate guardrails, env reads
@@ -750,220 +673,6 @@ async fn reborn_health_handler() -> Json<RebornHealthResponse> {
     })
 }
 
-// ─── auth middleware ──────────────────────────────────────────────────
-
-#[derive(Clone)]
-struct AuthLayerState {
-    tenant_id: TenantId,
-    default_agent_id: Option<AgentId>,
-    default_project_id: Option<ProjectId>,
-    authenticator: Arc<dyn WebuiAuthenticator>,
-    operator_routes: OperatorWebuiConfigRouteState,
-}
-
-/// Resolve `Authorization: Bearer <token>` for any v2 route, OR the
-/// `?token=…` query parameter only on the v2 SSE stream endpoint
-/// (mirrors the browser's `EventSource` limitation — it cannot set
-/// custom headers). On success, insert a [`WebUiAuthenticatedCaller`]
-/// extension built from the host-installation tenant + the
-/// authenticated user. On failure, return 401 before the v2 handler
-/// runs.
-async fn authenticate_request(
-    State(state): State<AuthLayerState>,
-    mut request: Request,
-    next: Next,
-) -> Response {
-    let token = match extract_bearer_token(&request) {
-        Some(token) => token,
-        None => return unauthorized(),
-    };
-
-    let auth = match state.authenticator.authenticate(&token).await {
-        Some(auth) => auth,
-        None => return unauthorized(),
-    };
-    if state
-        .operator_routes
-        .requires_operator_webui_config(&request)
-        && !auth.capabilities.operator_webui_config
-    {
-        return forbidden();
-    }
-
-    // Stamp the trusted agent/project from host installation config
-    // onto every authenticated caller. The downstream facade builds
-    // `ThreadScope` from `caller.agent_id` and 400s if it's missing,
-    // so a binary that fails to thread agent_id through here would
-    // authenticate users only to reject every v2 mutation/read. The
-    // browser body cannot influence either of these identifiers — by
-    // contract `WebuiServeConfig` is host-owned.
-    #[cfg(feature = "openai-compat-beta")]
-    let openai_user_id = auth.user_id.clone();
-    let caller = WebUiAuthenticatedCaller::new(
-        state.tenant_id.clone(),
-        auth.user_id,
-        state.default_agent_id.clone(),
-        state.default_project_id.clone(),
-    );
-    request.extensions_mut().insert(caller);
-    request.extensions_mut().insert(auth.capabilities);
-    #[cfg(feature = "openai-compat-beta")]
-    {
-        let scope = ironclaw_reborn_openai_compat::OpenAiCompatActorScope::new(
-            state.tenant_id.clone(),
-            openai_user_id.clone(),
-            state.default_agent_id.clone(),
-            state.default_project_id.clone(),
-        );
-        let auth_evidence = ironclaw_product_adapters::mark_bearer_token_verified_for_tenant(
-            openai_user_id.as_str(),
-            state.tenant_id.clone(),
-        );
-        let caller = match ironclaw_reborn_openai_compat::OpenAiCompatAuthenticatedCaller::new(
-            scope,
-            auth_evidence,
-        ) {
-            Ok(caller) => caller,
-            Err(_) => return unauthorized(),
-        };
-        request.extensions_mut().insert(caller);
-    }
-    next.run(request).await
-}
-
-fn unauthorized() -> Response {
-    (StatusCode::UNAUTHORIZED, "Invalid or missing auth token").into_response()
-}
-
-fn forbidden() -> Response {
-    (
-        StatusCode::FORBIDDEN,
-        "Operator WebUI configuration privileges required",
-    )
-        .into_response()
-}
-
-fn extract_bearer_token(request: &Request) -> Option<String> {
-    if let Some(value) = request.headers().get(header::AUTHORIZATION)
-        && let Ok(text) = value.to_str()
-        // `text.get(..7)` returns `None` when 7 is past the end OR
-        // lands inside a multi-byte UTF-8 sequence; both cases mean
-        // the value cannot be `Bearer <token>`. A direct byte slice
-        // would panic on a value whose first 7 bytes split a multi-byte
-        // character, which is forbidden for user-supplied data.
-        && let Some(prefix) = text.get(..7)
-        && prefix.eq_ignore_ascii_case("Bearer ")
-    {
-        // Safe: `prefix.eq_ignore_ascii_case("Bearer ")` matched, so
-        // the first 7 bytes are pure ASCII and byte 7 is a char
-        // boundary.
-        return Some(text[7..].to_string());
-    }
-    // `?token=` shim — only honored on the v2 SSE stream endpoint
-    // because `EventSource` cannot set request headers. Mutations and
-    // timeline reads stay bearer-only so a query-token leak in a
-    // referer chain cannot authenticate a state change.
-    //
-    // **Operational warning:** the token-as-URL-parameter pattern is
-    // a documented industry trade-off (SSE has no header-supplying
-    // client primitive). The token value appears in the URL and will
-    // therefore land in any HTTP access log, intermediate proxy log,
-    // or analytics pipeline that sees the request line. Composition
-    // emits `Referrer-Policy: no-referrer` on every response as
-    // defense in depth, but operators MUST still scrub
-    // `?token=<value>` from any log destination that retains URLs.
-    // The acceptance check is narrowed to GET on the exact
-    // `…/threads/{id}/events` path by `is_v2_sse_event_request` so
-    // the leak surface is one route, not the whole gateway.
-    if is_v2_sse_event_request(request) {
-        return query_token(request);
-    }
-    None
-}
-
-/// Returns `true` if the request is `GET /api/webchat/v2/threads/{id}/events`.
-/// The thread id must be a single non-empty path segment.
-pub(crate) fn is_v2_sse_event_request(request: &Request) -> bool {
-    if request.method() != Method::GET {
-        return false;
-    }
-    let path = request.uri().path();
-    let Some(rest) = path.strip_prefix("/api/webchat/v2/threads/") else {
-        return false;
-    };
-    let Some(thread_id) = rest.strip_suffix("/events") else {
-        return false;
-    };
-    !thread_id.is_empty() && !thread_id.contains('/')
-}
-
-fn query_token(request: &Request) -> Option<String> {
-    let query = request.uri().query()?;
-    url_query_value(query, "token")
-}
-
-fn url_query_value(query: &str, key: &str) -> Option<String> {
-    for pair in query.split('&') {
-        let mut parts = pair.splitn(2, '=');
-        let candidate_key = parts.next()?;
-        if candidate_key != key {
-            continue;
-        }
-        let raw_value = parts.next().unwrap_or("");
-        // Decode minimally: `+` → space, `%XX` → byte. Tokens are
-        // almost always opaque ASCII so we accept the value as-is and
-        // only handle the percent-decoded form when present. Empty or
-        // whitespace-only values count as absent so a stray `?token=`
-        // does not override a missing bearer header.
-        let decoded = percent_decode(raw_value);
-        let trimmed = decoded.trim();
-        if trimmed.is_empty() {
-            return None;
-        }
-        return Some(trimmed.to_string());
-    }
-    None
-}
-
-fn percent_decode(input: &str) -> String {
-    let bytes = input.as_bytes();
-    let mut out = Vec::with_capacity(bytes.len());
-    let mut i = 0;
-    while i < bytes.len() {
-        match bytes[i] {
-            b'+' => {
-                out.push(b' ');
-                i += 1;
-            }
-            b'%' if i + 2 < bytes.len() => {
-                let hi = hex_value(bytes[i + 1]);
-                let lo = hex_value(bytes[i + 2]);
-                if let (Some(hi), Some(lo)) = (hi, lo) {
-                    out.push((hi << 4) | lo);
-                    i += 3;
-                } else {
-                    out.push(bytes[i]);
-                    i += 1;
-                }
-            }
-            other => {
-                out.push(other);
-                i += 1;
-            }
-        }
-    }
-    String::from_utf8_lossy(&out).into_owned()
-}
-
-fn hex_value(b: u8) -> Option<u8> {
-    match b {
-        b'0'..=b'9' => Some(b - b'0'),
-        b'a'..=b'f' => Some(b - b'a' + 10),
-        b'A'..=b'F' => Some(b - b'A' + 10),
-        _ => None,
-    }
-}
-
 fn panic_handler(
     panic_info: Box<dyn std::any::Any + Send + 'static>,
 ) -> axum::http::Response<axum::body::Body> {
@@ -991,147 +700,4 @@ fn panic_handler(
         .unwrap_or_else(|_| {
             axum::http::Response::new(axum::body::Body::from("Internal Server Error"))
         })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::body::Body;
-    use axum::http::Method;
-
-    fn fake_request(method: Method, path_and_query: &str) -> Request {
-        Request::builder()
-            .method(method)
-            .uri(path_and_query)
-            .body(Body::empty())
-            .expect("request")
-    }
-
-    #[test]
-    fn v2_sse_event_request_recognized() {
-        let req = fake_request(Method::GET, "/api/webchat/v2/threads/abc/events");
-        assert!(is_v2_sse_event_request(&req));
-    }
-
-    #[test]
-    fn v2_sse_event_request_requires_get() {
-        let req = fake_request(Method::POST, "/api/webchat/v2/threads/abc/events");
-        assert!(!is_v2_sse_event_request(&req));
-    }
-
-    #[test]
-    fn v2_sse_event_request_requires_single_thread_segment() {
-        assert!(!is_v2_sse_event_request(&fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads//events"
-        )));
-        assert!(!is_v2_sse_event_request(&fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads/abc/events/extra"
-        )));
-    }
-
-    #[test]
-    fn v2_sse_event_request_rejects_other_v2_routes() {
-        assert!(!is_v2_sse_event_request(&fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads/abc/timeline"
-        )));
-        assert!(!is_v2_sse_event_request(&fake_request(
-            Method::POST,
-            "/api/webchat/v2/threads"
-        )));
-    }
-
-    #[test]
-    fn query_token_extracts_token_param() {
-        let req = fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads/abc/events?token=abc123",
-        );
-        assert_eq!(query_token(&req).as_deref(), Some("abc123"));
-    }
-
-    #[test]
-    fn query_token_decodes_percent_escapes() {
-        let req = fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads/abc/events?token=a%2Bb%3Dc",
-        );
-        assert_eq!(query_token(&req).as_deref(), Some("a+b=c"));
-    }
-
-    #[test]
-    fn query_token_treats_empty_as_absent() {
-        let req = fake_request(Method::GET, "/api/webchat/v2/threads/abc/events?token=");
-        assert!(query_token(&req).is_none());
-        let req2 = fake_request(
-            Method::GET,
-            "/api/webchat/v2/threads/abc/events?token=%20%20",
-        );
-        assert!(query_token(&req2).is_none());
-    }
-
-    #[test]
-    fn bearer_header_extraction_is_case_insensitive_on_prefix() {
-        let mut req = fake_request(Method::POST, "/api/webchat/v2/threads");
-        req.headers_mut().insert(
-            header::AUTHORIZATION,
-            HeaderValue::from_static("bEaReR mytoken"),
-        );
-        assert_eq!(extract_bearer_token(&req).as_deref(), Some("mytoken"));
-    }
-
-    #[test]
-    fn extract_bearer_token_rejects_query_token_on_non_sse_paths() {
-        // `?token=` is an EventSource-only escape hatch on the SSE
-        // route. Mutations and reads MUST stay bearer-only — a future
-        // regression that widens query-token acceptance to other
-        // routes would silently downgrade auth on every state change
-        // (no bearer header means an attacker only needs the URL).
-        // This test pins extract_bearer_token's behavior on every
-        // non-SSE shape we care about.
-        for (method, path_and_query) in [
-            (Method::POST, "/api/webchat/v2/threads?token=stealme"),
-            (
-                Method::POST,
-                "/api/webchat/v2/threads/abc/messages?token=stealme",
-            ),
-            (
-                Method::GET,
-                "/api/webchat/v2/threads/abc/timeline?token=stealme",
-            ),
-            (
-                Method::POST,
-                "/api/webchat/v2/threads/abc/runs/r/cancel?token=stealme",
-            ),
-            (
-                Method::POST,
-                "/api/webchat/v2/threads/abc/runs/r/gates/g/resolve?token=stealme",
-            ),
-            // Even on the SSE path, the wrong METHOD must reject.
-            (
-                Method::POST,
-                "/api/webchat/v2/threads/abc/events?token=stealme",
-            ),
-            // List threads shares the same path as create_thread but
-            // is read-only; query-token still rejected because no
-            // bearer header is present.
-            (Method::GET, "/api/webchat/v2/threads?token=stealme"),
-        ] {
-            let req = fake_request(method.clone(), path_and_query);
-            assert!(
-                extract_bearer_token(&req).is_none(),
-                "extract_bearer_token must NOT accept ?token= on {method} {path_and_query}",
-            );
-        }
-    }
-
-    #[test]
-    fn extract_bearer_token_accepts_query_token_only_on_sse_get() {
-        // Companion to the rejection test: the one place `?token=` is
-        // honored — GET on the SSE events route — must still work.
-        let req = fake_request(Method::GET, "/api/webchat/v2/threads/abc/events?token=ok");
-        assert_eq!(extract_bearer_token(&req).as_deref(), Some("ok"));
-    }
 }
